Moving changes to space point display to its own button
