+++ UserDev/EventDisplay/python/titus/modules/tpc_module.py
@@ -30,6 +30,7 @@ _DRAWABLE_LIST = {
     'End Point 2D': [drawables.EndPoint2D, "recob::EndPoint2D"],
     'Vertex': [drawables.Vertex, "recob::Vertex"],
     'Space Point': [drawables.SpacePoint, "recob::SpacePoint"],
+    'Space Point': [drawables.SpacePointHits, "recob::SpacePoint"]
 }
